V2/Backendv2/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

ultralytics==8.4.9
torch==2.5.1+cu121
torchvision==0.20.1+cu121
torchaudio==2.5.1+cu121

opencv-python==4.13.0.90
numpy==2.3.5
pandas==3.0.0
matplotlib==3.10.8
seaborn==0.13.2

scikit-learn==1.8.0
xgboost==3.1.3

fastapi==0.135.1
uvicorn==0.41.0

polars==1.37.1
PyYAML==6.0.3
tqdm==4.67.3
requests==2.32.5

python-multipart

streamlit
altair==4.2.2
requests

pillow-heif
plotly

joblib